webroms/requirements.txt
Django==2.1.2
djangorestframework==3.8.2
django-reversion==3.0.1
django-ratelimit==1.1.0
django-widget-tweaks==1.4.3
psycopg2==2.7.4
tzlocal==1.5.1